+++ examples/shirt-shop/utils/cart-types.ts
@@ -1,0 +1,9 @@
+export interface CartItem {
+  color: string;
+  size: string;
+  quantity: number;
+}
+
+export interface Cart {
+  items: CartItem[];
+}
